Add gitIgnore, Updated Sort method, Added URL printing
I changed the sorting so that Sort() is a single method that takes a string as a parameter. It then sorts the list of keys based on this parameter, e.g. 'pressure' or 'name'. I also added the URL printing to the print method inside keyswitch so that each key will automatically print the URL to the marketplace.